120 integrate starting set of validation tests (#139)
* Refactor structure of validation functions and linting
- Implement array for training validation functions: `from single_cellm.validation import TRAINING_VALIDATION_FUNCTION`
- `blacken` linting
* refactor tabula sapiens dataset generation
* Fix caching
* Expand cuda device selection function 
* Refactor validation functions to return a dict with reportable scores and a DataFrame
* Update functions.py: adapted docstring
---------

+++ src/single_cellm/utils/cuda.py
@@ -3,18 +3,29 @@ import torch
 import numpy as np
 
 
-def set_freest_gpu_as_device() -> int:
+def get_device() -> torch.device:
     """
     Sets the GPU with the most free memory as device, if available.
+
+    NOTE: This function is only relevant to Peter at the moment, who works on a machine with GPU sharing
+    NOTE: the use of this function might interfere with pytorch lightning, which itself has mechanisms to choose and use CPUs
     """
     if torch.cuda.is_available():
-        os.system("nvidia-smi -q -d Memory |grep -A4 GPU|grep Free > tmp_gpumem.txt")
-        memory_available = [
-            int(x.split()[2]) for x in open("tmp_gpumem.txt", "r").readlines()
-        ]
-        os.system("rm tmp_gpumem.txt")
-        device = torch.device(int(str(np.argmax(memory_available))))
-        torch.cuda.set_device(device)
-        print("Using GPU: {}".format(device))
+        if os.environ.get("HOSTNAME") == "011sv149":
+            os.system(
+                "nvidia-smi -q -d Memory |grep -A4 GPU|grep Free > tmp_gpumem.txt"
+            )
+            memory_available = [
+                int(x.split()[2]) for x in open("tmp_gpumem.txt", "r").readlines()
+            ]
+            os.system("rm tmp_gpumem.txt")
+            device = torch.device(int(str(np.argmax(memory_available))))
+            print("Using GPU: {}".format(device))
+        else:
+            device = torch.device("cuda")
+            print("Using GPU: {}".format(device))
     else:
-        print("No GPU available. Skipping setting freest GPU as device...")
+        print("No GPU available. Falling back to CPU")
+        device = torch.device("cpu")
+
+    return device
